+++ demo/lib/enable_hover_by_value_page.dart
@@ -18,11 +18,11 @@ class EnableHoverByValuePageState extends ExamplePageState {
 
   @override
   Widget buildContent() {
-    // coloring only the 'earth' feature
+    // coloring only the 'Darwin' feature
     MapChartTheme theme = MapChartTheme.value(
         key: 'Seq', colors: {4: Colors.green}, hoverColor: Colors.green[900]!);
 
-    // enabling hover only for the 'earth' feature
+    // enabling hover only for the 'Darwin' feature
     MapChart map = MapChart(
       dataSource: dataSource,
       theme: theme,